requirements.txt
# --- 1. CORE ML & DEPLOYMENT (Essential for Modal/FastAPI)
pandas
numpy
scikit-learn
xgboost
joblib
scipy

# API & Serverless
fastapi
uvicorn
pydantic
modal
python-dotenv

# Frontend (Streamlit)
streamlit

ipykernel
jupyter
apache-airflow
matplotlib
seaborn

requests






